BlueBytes/requirements.txt
fastapi>=0.116.0
python-docx>=1.1.0
python-multipart>=0.0.20
